--- Unity/PetEver/Assets/02.Scripts/MemorialUIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e895bdb7267c2bc4db27fba0bf1362e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
