+++ Unity/PetEver/Assets/02.Scripts/MemorySceneUIScript.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9febe54d51d991a419edcc785677e053
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
