+++ Unity/PetEver/Assets/02.Scripts/MemorySceneUIScript/UploadPicture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2461c24d14c0dfa45b75e59c72ea496b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
