--- Unity/PetEver/Assets/02.Scripts/UploadPicture.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2461c24d14c0dfa45b75e59c72ea496b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
