+++ Unity/PetEver/Assets/03.Prefabs/MemoryScene/NewpageUI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7eaa1ca773426544cb8609985e1f0fa9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
