+++ Unity/PetEver/Assets/03.Prefabs/MemoryScene/UserInpuuTextBGD.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 13a896cb0e91d6043aa9e2e34d639cbe
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
