+++ Unity/PetEver/Assets/04.Images/MemorialSceneUI/BGD.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3a95b954abaa6e6458197bafdd0e1927
+guid: a21498528f6cca845aeb10d828b8b542
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
