+++ Unity/PetEver/Assets/08.Fonts/notosanskr_bold SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be840354085997d41b3a86d51252bbbb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
